requirements.txt
requests
python-dotenv
speechrecognition
pyaudio
